+++ weblate/checks/models.py
@@ -65,6 +65,7 @@ class WeblateChecksConf(AppConf):
         "weblate.checks.chars.EndEllipsisCheck",
         "weblate.checks.chars.EndSemicolonCheck",
         "weblate.checks.chars.MaxLengthCheck",
+        "weblate.checks.chars.MultipleCapitalCheck",
         "weblate.checks.chars.KashidaCheck",
         "weblate.checks.chars.PunctuationSpacingCheck",
         "weblate.checks.chars.KabyleCharactersCheck",